+++ common/cpp/src/flutter_webrtc.cc
@@ -16,7 +16,8 @@ FlutterWebRTC::FlutterWebRTC(FlutterWebRTCPlugin* plugin)
       FlutterPeerConnection::FlutterPeerConnection(this),
       FlutterScreenCapture::FlutterScreenCapture(this),
       FlutterDataChannel::FlutterDataChannel(this),
-      FlutterFrameCryptor::FlutterFrameCryptor(this) {}
+      FlutterFrameCryptor::FlutterFrameCryptor(this),
+      FlutterDataPacketCryptor::FlutterDataPacketCryptor(this) {}
 
 FlutterWebRTC::~FlutterWebRTC() {}
 
@@ -1289,6 +1290,9 @@ void FlutterWebRTC::HandleMethodCall(
   } else {
     if (HandleFrameCryptorMethodCall(method_call, std::move(result), &result)) {
       return;
+    } else if (HandleDataPacketCryptorMethodCall(method_call, std::move(result),
+                                                 &result)) {
+      return;
     } else {
       result->NotImplemented();
     }
@@ -1296,11 +1300,11 @@ void FlutterWebRTC::HandleMethodCall(
 }
 
 void FlutterWebRTC::initLoggerCallback(RTCLoggingSeverity severity) {
-  if(eventChannelProxy == nullptr) {
+  if (eventChannelProxy == nullptr) {
     eventChannelProxy = event_channel();
   }
 
-  libwebrtc::LibWebRTCLogging::setLogSink(severity, [](const string& message){
+  libwebrtc::LibWebRTCLogging::setLogSink(severity, [](const string& message) {
     EncodableMap info;
     info[EncodableValue("event")] = "onLogData";
     info[EncodableValue("data")] = message.c_string();
@@ -1309,15 +1313,15 @@ void FlutterWebRTC::initLoggerCallback(RTCLoggingSeverity severity) {
 }
 
 RTCLoggingSeverity FlutterWebRTC::str2LogSeverity(std::string str) {
-  if(str == "verbose")
+  if (str == "verbose")
     return Verbose;
-  else if(str == "info")
+  else if (str == "info")
     return Info;
-  else if(str == "warning")
+  else if (str == "warning")
     return Warning;
-  else if(str == "error")
+  else if (str == "error")
     return Error;
-  else if(str == "none")
+  else if (str == "none")
     return None;
 
   return None;